backend/requirements.txt
asgiref==3.7.2
astroid==2.15.6
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.2.2
cloudpickle==3.0.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
ec2-hibinit-agent==1.0.0
fonttools==4.43.1
gyp==0.1
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isort==5.12.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
llvmlite==0.41.1
MarkupSafe==2.0.1
matplotlib==3.8.0
mccabe==0.7.0
more-itertools==8.10.0
mysql-connector-python==8.1.0
mysqlclient==2.1.1
netifaces==0.11.0
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.0
packaging==23.2
pandas==2.1.1
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.10.0
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.17.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==1.3.1
scipy==1.11.3
SecretStorage==3.3.1
service-identity==18.1.0
shap==0.43.0
six==1.16.0
slicer==0.0.7
sos==4.5.6
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
Twisted==22.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wrapt==1.15.0
xgboost==2.0.0
zipp==1.0.0
zope.interface==5.4.0
